Separate logger tests for each model
Just too many tests for single file.
