+++ geokey/core/tests/logger/test_log_lookupfield.py
@@ -1,0 +1,200 @@
+"""Tests for logger: model LookupField."""
+
+from django.test import TestCase
+
+from geokey.core.models import LoggerHistory
+from geokey.users.tests.model_factories import UserFactory
+from geokey.projects.tests.model_factories import ProjectFactory
+from geokey.categories.tests.model_factories import (
+    CategoryFactory,
+    LookupFieldFactory,
+)
+
+
+class LogLookupFieldTest(TestCase):
+    """Test model LookupField."""
+
+    def setUp(self):
+        """Set up test."""
+        self.user = UserFactory.create()
+        self.project = ProjectFactory.create(**{
+            'creator': self.user})
+        self.category = CategoryFactory.create(**{
+            'creator': self.user,
+            'project': self.project})
+        self.lookupfield = LookupFieldFactory.create(**{
+            'category': self.category})
+
+    def test_log_create(self):
+        """Test when lookup field gets created."""
+        log_count_init = LoggerHistory.objects.count()
+        field = LookupFieldFactory.create(**{
+            'category': self.category})
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(field.id),
+            'name': field.name,
+            'type': 'LookupField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'created'})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+    def test_log_update_name(self):
+        """Test when name changes."""
+        log_count_init = LoggerHistory.objects.count()
+        self.lookupfield.name = '%s UPDATED' % self.lookupfield.name
+        self.lookupfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.lookupfield.id),
+            'name': self.lookupfield.name,
+            'type': 'LookupField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'name',
+            'value': self.lookupfield.name})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+    def test_log_update_status(self):
+        """Test when  status changes."""
+        log_count_init = LoggerHistory.objects.count()
+
+        self.lookupfield.status = 'inactive'
+        self.lookupfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.lookupfield.id),
+            'name': self.lookupfield.name,
+            'type': 'LookupField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'status',
+            'value': self.lookupfield.status})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+        self.lookupfield.status = 'active'
+        self.lookupfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.lookupfield.id),
+            'name': self.lookupfield.name,
+            'type': 'LookupField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'status',
+            'value': self.lookupfield.status})
+        self.assertEqual(log_count, log_count_init + 2)
+        self.assertEqual(log.historical, None)
+
+    def test_log_update_required(self):
+        """Test when required setting changes."""
+        log_count_init = LoggerHistory.objects.count()
+
+        self.lookupfield.required = True
+        self.lookupfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.lookupfield.id),
+            'name': self.lookupfield.name,
+            'type': 'LookupField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'required',
+            'value': str(self.lookupfield.required)})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+        self.lookupfield.required = False
+        self.lookupfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.lookupfield.id),
+            'name': self.lookupfield.name,
+            'type': 'LookupField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'required',
+            'value': str(self.lookupfield.required)})
+        self.assertEqual(log_count, log_count_init + 2)
+        self.assertEqual(log.historical, None)
