+++ geokey/core/tests/logger/test_log_numericfield.py
@@ -1,0 +1,200 @@
+"""Tests for logger: model NumericField."""
+
+from django.test import TestCase
+
+from geokey.core.models import LoggerHistory
+from geokey.users.tests.model_factories import UserFactory
+from geokey.projects.tests.model_factories import ProjectFactory
+from geokey.categories.tests.model_factories import (
+    CategoryFactory,
+    NumericFieldFactory,
+)
+
+
+class LogNumericFieldTest(TestCase):
+    """Test model NumericField."""
+
+    def setUp(self):
+        """Set up test."""
+        self.user = UserFactory.create()
+        self.project = ProjectFactory.create(**{
+            'creator': self.user})
+        self.category = CategoryFactory.create(**{
+            'creator': self.user,
+            'project': self.project})
+        self.numericfield = NumericFieldFactory.create(**{
+            'category': self.category})
+
+    def test_log_create(self):
+        """Test when numeric field gets created."""
+        log_count_init = LoggerHistory.objects.count()
+        field = NumericFieldFactory.create(**{
+            'category': self.category})
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(field.id),
+            'name': field.name,
+            'type': 'NumericField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'created'})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+    def test_log_update_name(self):
+        """Test when name changes."""
+        log_count_init = LoggerHistory.objects.count()
+        self.numericfield.name = '%s UPDATED' % self.numericfield.name
+        self.numericfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.numericfield.id),
+            'name': self.numericfield.name,
+            'type': 'NumericField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'name',
+            'value': self.numericfield.name})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+    def test_log_update_status(self):
+        """Test when status changes."""
+        log_count_init = LoggerHistory.objects.count()
+
+        self.numericfield.status = 'inactive'
+        self.numericfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.numericfield.id),
+            'name': self.numericfield.name,
+            'type': 'NumericField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'status',
+            'value': self.numericfield.status})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+        self.numericfield.status = 'active'
+        self.numericfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.numericfield.id),
+            'name': self.numericfield.name,
+            'type': 'NumericField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'status',
+            'value': self.numericfield.status})
+        self.assertEqual(log_count, log_count_init + 2)
+        self.assertEqual(log.historical, None)
+
+    def test_log_update_required(self):
+        """Test when required setting changes."""
+        log_count_init = LoggerHistory.objects.count()
+
+        self.numericfield.required = True
+        self.numericfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.numericfield.id),
+            'name': self.numericfield.name,
+            'type': 'NumericField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'required',
+            'value': str(self.numericfield.required)})
+        self.assertEqual(log_count, log_count_init + 1)
+        self.assertEqual(log.historical, None)
+
+        self.numericfield.required = False
+        self.numericfield.save()
+
+        log = LoggerHistory.objects.last()
+        log_count = LoggerHistory.objects.count()
+
+        self.assertNotEqual(log.user, {
+            'id': str(self.user.id),
+            'display_name': self.user.display_name})
+        self.assertEqual(log.project, {
+            'id': str(self.project.id),
+            'name': self.project.name})
+        self.assertEqual(log.category, {
+            'id': str(self.category.id),
+            'name': self.category.name})
+        self.assertEqual(log.field, {
+            'id': str(self.numericfield.id),
+            'name': self.numericfield.name,
+            'type': 'NumericField'})
+        self.assertEqual(log.subset, None)
+        self.assertEqual(log.action, {
+            'id': 'updated',
+            'field': 'required',
+            'value': str(self.numericfield.required)})
+        self.assertEqual(log_count, log_count_init + 2)
+        self.assertEqual(log.historical, None)
